refactor: split accessible slide composer
